ena.4: optimize apropos and hardware
Remove "FreeBSD kernel driver for" and "family", and
add "AWS EC2" and "driver" to the document description
for better search discoveribility and consistency with
other driver manuals.

Rewrite the introductory sentance to the HARDWARE section
with driver name for inclusion in hardware release notes.

MFC after:	3 days
Reviewed by:	<osamaabb@amazon.com>
Sponsored by:	Google