sysexits.3: Enable more robust apropos
Markup exit status codes as defined variables. This produces no change
in rendered output, however enables this workflow:

$ apropos Dv=67
sysexits(3) - legacy exit status codes for system programs

Markup error constants as error constants. This unfortunately removes
bold from rendered output, but it enables:

$ apropos Er=EX_USAGE
sysexits(3) - legacy exit status codes for system programs

This moves to align with how the mdoc(7) language was designed.
IMHO we should campaign to get errors rendered in bold instead of not
using the language. On the community discord, kevans mused that this
could be useful for identifying where the problem is in a complex
script.

MFC after:	3 days