build/development.7: System building examples
The manual is notably missing building a custom kernel. To polish the
idea of a complete, well-integrated system, the integrated doc could
use this. Upon reviewing development(7), the description says that
"build instructions can be found in build(7)", which makes a lot of
sense to me.

Therefore, move all of these examples on building the system to build(7)
and bring the instructions up to date with src/UPDATING. Provide a new
example for building a custom kernel.

While here, give build(7)/FILES some love, and the example for
PORTS_MODULES was indented far enough to wrap on std. console. Set that
back to increase legibility.

Outstanding:
- development(7) needs a section on github
- it could probably use subsections
- build(7) has potentially confusing language in the introduction

MFC after:		3 days
Reported by:		emaste (manpages needs a kernconf example)
Reported by:		imp (kernconfs include and go from there)