build/development.7: System building examples
Consolidate examples to build the system in the build(7) manual where
the details about the syntax involved is documented. Remove them from
development(7), instead adding examples for patching the trees.

Reduce documentation proliferation and create consistent, maintainable
doc about building the system, eventually removing the COMMMON ITEMS
section of UPDATING so there is a single source of truth in the cannon
and well-integrated location, the in-band system reference manual.

This commit refactors the introduction of build(7) for presentation and
introduces six detailed examples, mostly from development(7) and
UPDATING, for building the system.

+ Example 1: Build and upgrade system in place
+ Example 2: Build and upgrade a custom kernel in place
+ Example 3: Build and upgrade a single piece of userspace
+ Example 4: Build and upgrade a loadable kernel module
+ Example 5: Quickly rebuild a kernel in place
+ Example 6: Cross-compiling for different architectures

Add a diagnostics section redirecting the mistaken to these examples.

MFC after:		3 days
Discussed with:		emaste, imp, jhb, ivy
Differential Revision:	https://reviews.freebsd.org/D48693